+++ R/method-ggplot-add.R
@@ -107,6 +107,18 @@ ggplot_add.cladelabel <- function(object, plot, object_name) {
     ggplot_add(ly, plot, object_name)
 }
 
+##' @method ggplot_add hilight
+##' @export
+ggplot_add.hilight <- function(object, plot, object_name) {
+    layout <- get("layout", envir = plot$plot_env)
+    if (layout == "unrooted" || layout == "daylight") {
+        ly <- do.call(geom_hilight_encircle, object)
+    } else {
+        ly <- do.call(geom_hilight_rectangular, object)
+    }
+    ggplot_add(ly, plot, object_name)
+}
+
 ##' @method ggplot_add striplabel
 ##' @export
 ggplot_add.striplabel <- function(object, plot, object_name) {